# some performance optimizations (especially XML parsing)