--- a/src/openms/include/OpenMS/FORMAT/Base64.h
+++ b/src/openms/include/OpenMS/FORMAT/Base64.h
@@ -67,10 +67,7 @@ namespace OpenMS
 public:
 
     /// default constructor
-    Base64();
-
-    /// Destructor
-    virtual ~Base64();
+    Base64() = default;
 
     /// Byte order type
     enum ByteOrder
@@ -87,7 +84,7 @@ public:
         @note @p in will be empty after this method
     */
     template <typename FromType>
-    void encode(std::vector<FromType> & in, ByteOrder to_byte_order, String & out, bool zlib_compression = false);
+    static void encode(std::vector<FromType> & in, ByteOrder to_byte_order, String & out, bool zlib_compression = false);
 
     /**
         @brief Decodes a Base64 string to a vector of floating point numbers
@@ -95,7 +92,7 @@ public:
         You have to specify the byte order of the input and if it is zlib-compressed.
     */
     template <typename ToType>
-    void decode(const String & in, ByteOrder from_byte_order, std::vector<ToType> & out, bool zlib_compression = false);
+    static void decode(const String & in, ByteOrder from_byte_order, std::vector<ToType> & out, bool zlib_compression = false);
 
     /**
         @brief Encodes a vector of integer point numbers to a Base64 string
@@ -105,7 +102,7 @@ public:
         @note @p in will be empty after this method
     */
     template <typename FromType>
-    void encodeIntegers(std::vector<FromType> & in, ByteOrder to_byte_order, String & out, bool zlib_compression = false);
+    static void encodeIntegers(std::vector<FromType> & in, ByteOrder to_byte_order, String & out, bool zlib_compression = false);
 
     /**
         @brief Decodes a Base64 string to a vector of integer numbers
@@ -113,7 +110,7 @@ public:
         You have to specify the byte order of the input and if it is zlib-compressed.
     */
     template <typename ToType>
-    void decodeIntegers(const String & in, ByteOrder from_byte_order, std::vector<ToType> & out, bool zlib_compression = false);
+    static void decodeIntegers(const String & in, ByteOrder from_byte_order, std::vector<ToType> & out, bool zlib_compression = false);
 
     /**
         @brief Encodes a vector of strings to a Base64 string
@@ -127,7 +124,7 @@ public:
       
         @note Unless append_null_byte is false, will add a null byte ("\0") at the end of each input
     */
-    void encodeStrings(const std::vector<String> & in, String & out, bool zlib_compression = false, bool append_null_byte = true);
+    static void encodeStrings(const std::vector<String> & in, String & out, bool zlib_compression = false, bool append_null_byte = true);
 
     /**
         @brief Decodes a Base64 string to a vector of (null-terminated) strings
@@ -138,7 +135,7 @@ public:
         @param out A vector containing the decoded data (split at null "\0") bytes
         @param zlib_compression Whether the data should be decompressed with zlib after decoding in Base64
     */
-    void decodeStrings(const String & in, std::vector<String> & out, bool zlib_compression = false);
+    static void decodeStrings(const String & in, std::vector<String> & out, bool zlib_compression = false);
 
     /**
         @brief Decodes a Base64 string to a QByteArray
@@ -147,7 +144,7 @@ public:
         @param out A ByteArray containing the decoded data
         @param zlib_compression Whether the data should be decompressed with zlib after decoding in Base64
     */
-    void decodeSingleString(const String & in, QByteArray & base64_uncompressed, bool zlib_compression);
+    static void decodeSingleString(const String & in, QByteArray & base64_uncompressed, bool zlib_compression);
 
 private:
 
@@ -169,19 +166,19 @@ private:
     static const char decoder_[];
     /// Decodes a Base64 string to a vector of floating point numbers
     template <typename ToType>
-    void decodeUncompressed_(const String & in, ByteOrder from_byte_order, std::vector<ToType> & out);
+    static void decodeUncompressed_(const String & in, ByteOrder from_byte_order, std::vector<ToType> & out);
 
     ///Decodes a compressed Base64 string to a vector of floating point numbers
     template <typename ToType>
-    void decodeCompressed_(const String & in, ByteOrder from_byte_order, std::vector<ToType> & out);
+    static void decodeCompressed_(const String & in, ByteOrder from_byte_order, std::vector<ToType> & out);
 
     /// Decodes a Base64 string to a vector of integer numbers
     template <typename ToType>
-    void decodeIntegersUncompressed_(const String & in, ByteOrder from_byte_order, std::vector<ToType> & out);
+    static void decodeIntegersUncompressed_(const String & in, ByteOrder from_byte_order, std::vector<ToType> & out);
 
     ///Decodes a compressed Base64 string to a vector of integer numbers
     template <typename ToType>
-    void decodeIntegersCompressed_(const String & in, ByteOrder from_byte_order, std::vector<ToType> & out);
+    static void decodeIntegersCompressed_(const String & in, ByteOrder from_byte_order, std::vector<ToType> & out);
   };
 
   /// Endianizes a 32 bit type from big endian to little endian and vice versa