--- a/src/openms/include/OpenMS/FORMAT/HANDLERS/MzMLHandlerHelper.h
+++ b/src/openms/include/OpenMS/FORMAT/HANDLERS/MzMLHandlerHelper.h
@@ -57,33 +57,34 @@ namespace OpenMS
       /// Binary data representation
       struct BinaryData
       {
-        String base64;
+        // ordered by size (alignment) and cache hotness in 'decode'
         enum {PRE_NONE, PRE_32, PRE_64} precision;
-        Size size;
+        enum { DT_NONE, DT_FLOAT, DT_INT, DT_STRING } data_type;
+        MSNumpressCoder::NumpressCompression np_compression;
         bool compression; // zlib compression
-        enum {DT_NONE, DT_FLOAT, DT_INT, DT_STRING} data_type;
+        String base64;
+        Size size;
         std::vector<float> floats_32;
         std::vector<double> floats_64;
         std::vector<Int32> ints_32;
         std::vector<Int64> ints_64;
         std::vector<String> decoded_char;
         MetaInfoDescription meta;
-        MSNumpressCoder::NumpressCompression np_compression;
 
         /// Constructor
         BinaryData() :
-          base64(),
           precision(PRE_NONE),
-          size(0),
-          compression(false),
           data_type(DT_NONE),
+          np_compression(),
+          compression(false),
+          base64(),
+          size(0),
           floats_32(),
           floats_64(),
           ints_32(),
           ints_64(),
           decoded_char(),
-          meta(),
-          np_compression()
+          meta()
         {
         }
 
@@ -110,7 +111,7 @@ namespace OpenMS
       */
       static void decodeBase64Arrays(std::vector<BinaryData> & data_, bool skipXMLCheck = false);
 
-      static void computeDataProperties_(std::vector<BinaryData>& data_, bool& precision_64, SignedSize& index, String index_name);
+      static void computeDataProperties_(std::vector<BinaryData>& data_, bool& precision_64, SignedSize& index, const String& index_name);
 
       static bool handleBinaryDataArrayCVParam(std::vector<BinaryData>& data_,
         const String& accession, const String& value, const String& name);