[FEATURE] faster whitespace removal function (2x - 6x faster than before)

On string with no whitespaces (stringlen 3000, 1e5 iterations):
old: 4.46 s (wall), 4.46 s (CPU), 0.00 s (system), 4.46 s (user)
new: 2.68 s (wall), 2.68 s (CPU), 0.00 s (system), 2.68 s (user)

On string with whitespaces: (stringlen 3000, 1e5 iterations):
old: 25.03 s (wall), 25.02 s (CPU), 0.00 s (system), 25.02 s (user)
new: 3.37 s (wall), 3.37 s (CPU), 0.00 s (system), 3.37 s (user)

[FIX] some cleaning up in XML parsing code, e.g. Base64 uses static functions only, so we can remove instanciation

In total, mzML parsing (FileInfo -in ...) is about 6% faster (~60 vs 63 MB/sec) on my machine.
There is a lot more potential IMHO, but I take what I can get :)