@timosachsenberg 2018-02-05
looks good. some minor style issues on CDASH http://cdash.openms.de/testDetails.php?test=47239110&build=256751
Others are probably related to the travis cache not being cleared...