@hroest 2018-02-06
👍 if the tests passes, some comments in the code