@hroest 2018-02-06
did you compare this to `str.erase(remove_if(str.begin(), str.end(), isspace), str.end());` ?