@cbielow 2018-02-06
Not before, but our own implementation seems to do way better:

Without spaces:

Erase: 6.25 s (wall), 6.24 s (CPU), 0.00 s (system), 6.24 s (user)
this PR: 2.69 s (wall), 2.68 s (CPU), 0.00 s (system), 2.68 s (user)

With spaces:
Erase: 7.63 s (wall), 7.61 s (CPU), 0.00 s (system), 7.61 s (user)
this PR: 3.68 s (wall), 3.67 s (CPU), 0.00 s (system), 3.67 s (user)


From: Hannes Roest [mailto:notifications@github.com]
Sent: Tuesday, February 06, 2018 1:23 AM
To: OpenMS/OpenMS <OpenMS@noreply.github.com>
Cc: Bielow, Chris <Chris.Bielow@fu-berlin.de>; Assign <assign@noreply.github.com>
Subject: Re: [OpenMS/OpenMS] some performance optimizations (especially XML parsing) (#3223)


did you compare this to str.erase(remove_if(str.begin(), str.end(), isspace), str.end()); ?

—
You are receiving this because you were assigned.
Reply to this email directly, view it on GitHub<https://github.com/OpenMS/OpenMS/pull/3223#issuecomment-363267072>, or mute the thread<https://github.com/notifications/unsubscribe-auth/AFuvkoiNW7X23l9uRbs3_VoAD1DWYI09ks5tR5tegaJpZM4R5RNU>.