@cbielow 2018-02-07
tests look good (except for some false positives). will merge