@hroest 2018-02-07
> remove mem test after free (not guaranteed to work on Linux) 2

hmm not sure I agree, why was this working before and now suddenly it doesnt work any more? sounds suspicious ?