@cbielow 2018-02-08
I've been digging through documentation on how (virtual-)memory/paging is handled in C++ libs, but did not really get a good answer.
All I can say is: I'm very sure there is no (new) mem-leak since the code change is minimal and when used on Windows, you get all the memory back (even when iterating over loading/releasing an MSExperiment 100 times in the SysInfo test, the memory after the last release is the same as before the first).
So assuming that g++/clang does not have a buggy memory manager, the only explanation which is left is that upon 'free()' (if used under the hood of STL's delete operator), the memory manager decides not to return memory (yet) to the OS but rather hold onto it for the next malloc/new.
(see https://www.gnu.org/software/libc/manual/html_node/Freeing-after-Malloc.html#Freeing-after-Malloc).
I don't see why this would be a bad thing.